initial chain name as url refactoring
