+++ examples/local-explorer/mprocs.yaml
@@ -18,4 +18,4 @@ procs:
       SQLITE_FILENAME: "indexer.db"
   explorer:
     cwd: packages/contracts
-    shell: pnpm explorer --dev
+    shell: pnpm explorer
