+++ packages/explorer/src/app/[chainName]/(explorer)/worlds/[worldAddress]/explore/EditableTableCell.tsx
@@ -9,11 +9,11 @@ import { SchemaAbiType } from "@latticexyz/schema-type/internal";
 import IBaseWorldAbi from "@latticexyz/world/out/IBaseWorld.sol/IBaseWorld.abi.json";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { waitForTransactionReceipt, writeContract } from "@wagmi/core";
-import { getChain } from "../../../../../common";
-import { Checkbox } from "../../../../../components/ui/Checkbox";
-import { camelCase, cn } from "../../../../../lib/utils";
-import { TableConfig } from "../../../../api/table/route";
-import { wagmiConfig } from "../../../Providers";
+import { getChain } from "../../../../../../common";
+import { Checkbox } from "../../../../../../components/ui/Checkbox";
+import { camelCase, cn } from "../../../../../../lib/utils";
+import { TableConfig } from "../../../../../api/table/route";
+import { wagmiConfig } from "../../../../Providers";
 
 type Props = {
   name: string;
