+++ packages/explorer/src/app/[chainName]/(explorer)/worlds/[worldAddress]/explore/page.tsx
