+++ packages/explorer/src/app/[chainName]/(explorer)/worlds/[worldAddress]/interact/FunctionField.tsx
@@ -6,10 +6,10 @@ import { z } from "zod";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Button } from "../../../../../components/ui/Button";
-import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "../../../../../components/ui/Form";
-import { Input } from "../../../../../components/ui/Input";
-import { Separator } from "../../../../../components/ui/Separator";
+import { Button } from "../../../../../../components/ui/Button";
+import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "../../../../../../components/ui/Form";
+import { Input } from "../../../../../../components/ui/Input";
+import { Separator } from "../../../../../../components/ui/Separator";
 import { useContractMutation } from "./useContractMutation";
 
 export enum FunctionType {
