+++ packages/explorer/src/app/[chainName]/(explorer)/worlds/[worldAddress]/interact/useContractMutation.ts
@@ -4,8 +4,8 @@ import { Abi, AbiFunction, Hex } from "viem";
 import { useAccount } from "wagmi";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { readContract, waitForTransactionReceipt, writeContract } from "@wagmi/core";
-import { getChain } from "../../../../../common";
-import { wagmiConfig } from "../../../Providers";
+import { getChain } from "../../../../../../common";
+import { wagmiConfig } from "../../../../Providers";
 import { FunctionType } from "./FunctionField";
 
 type UseContractMutationProps = {
@@ -17,8 +17,8 @@ const chain = getChain();
 const chainId = chain.id;
 
 export function useContractMutation({ abi, operationType }: UseContractMutationProps) {
-  const queryClient = useQueryClient();
   const { worldAddress } = useParams();
+  const queryClient = useQueryClient();
   const account = useAccount();
 
   return useMutation({
