+++ packages/explorer/src/app/[chainName]/(explorer)/worlds/[worldAddress]/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Navigation } from "../../../../components/Navigation";
+import { Navigation } from "../../../../../components/Navigation";
 
 export default function WorldLayout({ children }: { children: React.ReactNode }) {
   return (
