+++ packages/explorer/src/app/[chainName]/Providers.tsx
@@ -1,12 +1,13 @@
 "use client";
 
+import { useParams } from "next/navigation";
 import { WagmiProvider, createConfig, http } from "wagmi";
 import { injected, metaMask, safe } from "wagmi/connectors";
 import { ReactNode } from "react";
 import { RainbowKitProvider, darkTheme } from "@rainbow-me/rainbowkit";
 import "@rainbow-me/rainbowkit/styles.css";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { getChain } from "../../common";
+import { getChain, namedChains } from "../../common";
 import { defaultAnvilConnectors } from "../../connectors/anvil";
 
 const queryClient = new QueryClient();
@@ -31,6 +32,30 @@ export const wagmiConfig = createConfig({
 });
 
 export function Providers({ children }: { children: ReactNode }) {
+  const { chainName } = useParams();
+  const chain = namedChains[chainName as string];
+  if (!chain) {
+    throw new Error(`Chain ${chainName} not supported`);
+  }
+
+  const wagmiConfig = createConfig({
+    chains: [chain],
+    connectors: [
+      injected(),
+      metaMask({
+        dappMetadata: {
+          name: "World Explorer",
+        },
+      }),
+      safe(),
+      ...defaultAnvilConnectors,
+    ],
+    transports: {
+      [chain.id]: http(),
+    },
+    ssr: true,
+  });
+
   return (
     <WagmiProvider config={wagmiConfig}>
       <QueryClientProvider client={queryClient}>
