+++ packages/explorer/src/app/[chainName]/error.tsx
