+++ packages/explorer/src/app/[chainName]/layout.tsx
